Buy Order
- Security selecteren
- Orderformulier tonen
- Gebruiker moet formulier in kunnen vullen
- Ordertype kiezen
- Extra requirements bij limit order
- Totaalprijs en fees berekenen en tonen
- Onvoldoende cash melding
- Order samenvatting
- Order bevestigen
- Cash reserven
- Order aanmaken
- Ordernummen generen
- Order naar beurs sturen
- Holdings bijwerken na uitvoering order
- Reservatie definitief afboeken
- Conformatie melding

+++ app/DTO/PortfolioDTO.php
@@ -11,7 +11,7 @@ class PortfolioDTO extends DTO
 {
     public int $id;
 
-    public float $cash;
+    public int $cash;
 
     public float $total_value;
